requirements.txt
-i https://pypi.org/simple
appdirs==1.4.3
bokeh==1.3.4
certifi==2019.9.11
chardet==3.0.4
click==7.0
cloudpickle==1.2.2
dask[bag,dataframe]==2.5.0
dominate==2.4.0
entrypoints==0.3
flask-bootstrap4==4.0.2
flask-wtf==0.14.2
flask==1.1.1
fsspec==0.5.1
holoviews==1.12.5
hvplot==0.4.0
idna==2.8
intake[complete]==0.5.3
itsdangerous==1.1.0
jinja2==2.10.1
locket==0.2.0
markdown==3.1.1
markupsafe==1.1.1
msgpack-numpy==0.4.4.3
msgpack==0.6.2
numpy==1.17.2
packaging==19.2
pandas==0.25.1
panel==0.6.3
param==1.9.1
partd==1.0.0
pillow==6.1.0
pyct==0.4.6
pyparsing==2.4.2
python-dateutil==2.8.0
python-snappy==0.5.4
pytz==2019.2
pyviz-comms==0.7.2
pyyaml==5.1.2
requests==2.22.0
six==1.12.0
toolz==0.10.0
tornado==6.0.3
urllib3==1.25.6
uwsgi==2.0.18
visitor==0.1.3
werkzeug==0.16.0
wtforms==2.2.1
